Move document-state to services
