--- src/document-state.js
@@ -1,155 +1,0 @@
-/**
- * Copyright 2015 The AMP HTML Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS-IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import {Observable} from './observable';
-import {fromClass} from './service';
-import {getVendorJsPropertyName} from './style';
-import {waitForChild} from './dom';
-
-
-/**
- */
-export class DocumentState {
-  /**
-   * @param {!Window} win
-   */
-  constructor(win) {
-    /** @const {!Window} */
-    this.win = win;
-
-    /** @private @const {!Document} */
-    this.document_ = win.document;
-
-    /** @private {string|null} */
-    this.hiddenProp_ = getVendorJsPropertyName(this.document_, 'hidden', true);
-    if (this.document_[this.hiddenProp_] === undefined) {
-      this.hiddenProp_ = null;
-    }
-
-    /** @private {string|null} */
-    this.visibilityStateProp_ = getVendorJsPropertyName(this.document_,
-        'visibilityState', true);
-    if (this.document_[this.visibilityStateProp_] === undefined) {
-      this.visibilityStateProp_ = null;
-    }
-
-    /** @private @const {!Observable} */
-    this.visibilityObservable_ = new Observable();
-
-    /** @private {string|null} */
-    this.visibilityChangeEvent_ = null;
-    if (this.hiddenProp_) {
-      this.visibilityChangeEvent_ = 'visibilitychange';
-      const vendorStop = this.hiddenProp_.indexOf('Hidden');
-      if (vendorStop != -1) {
-        this.visibilityChangeEvent_ =
-            this.hiddenProp_.substring(0, vendorStop) + 'Visibilitychange';
-      }
-    }
-
-    /** @private @const {!Function} */
-    this.boundOnVisibilityChanged_ = this.onVisibilityChanged_.bind(this);
-    if (this.visibilityChangeEvent_) {
-      this.document_.addEventListener(this.visibilityChangeEvent_,
-          this.boundOnVisibilityChanged_);
-    }
-
-    /** @private {?Observable} */
-    this.bodyAvailableObservable_ = null;
-  }
-
-  /** @private */
-  cleanup_() {
-    if (this.visibilityChangeEvent_) {
-      this.document_.removeEventListener(this.visibilityChangeEvent_,
-          this.boundOnVisibilityChanged_);
-    }
-  }
-
-  /**
-   * Returns the value of "document.hidden" property. The reasons why it may
-   * not be visible include document in a non-active tab or when the document
-   * is being pre-rendered via link with rel="prerender".
-   * @return {boolean}
-   */
-  isHidden() {
-    if (!this.hiddenProp_) {
-      return false;
-    }
-    return this.document_[this.hiddenProp_];
-  }
-
-  /**
-   * Returns the value of "document.visibilityState" property. Possible values
-   * are: 'hidden', 'visible', 'prerender', and 'unloaded'.
-   * @return {string}
-   */
-  getVisibilityState() {
-    if (!this.visibilityStateProp_) {
-      return this.isHidden() ? 'hidden' : 'visible';
-    }
-    return this.document_[this.visibilityStateProp_];
-  }
-
-  /**
-   * @param {function()} handler
-   * @return {!UnlistenDef}
-   */
-  onVisibilityChanged(handler) {
-    return this.visibilityObservable_.add(handler);
-  }
-
-  /** @private */
-  onVisibilityChanged_() {
-    this.visibilityObservable_.fire();
-  }
-
-  /**
-   * If body is already available, callback is called synchronously and null
-   * is returned.
-   * @param {function()} handler
-   * @return {?UnlistenDef}
-   */
-  onBodyAvailable(handler) {
-    const doc = this.document_;
-    if (doc.body) {
-      handler();
-      return null;
-    }
-    if (!this.bodyAvailableObservable_) {
-      this.bodyAvailableObservable_ = new Observable();
-      waitForChild(doc.documentElement, () => !!doc.body,
-          this.onBodyAvailable_.bind(this));
-    }
-    return this.bodyAvailableObservable_.add(handler);
-  }
-
-  /** @private */
-  onBodyAvailable_() {
-    this.bodyAvailableObservable_.fire();
-    this.bodyAvailableObservable_.removeAll();
-    this.bodyAvailableObservable_ = null;
-  }
-}
-
-
-/**
- * @param {!Window} window
- * @return {!DocumentState}
- */
-export function documentStateFor(window) {
-  return fromClass(window, 'documentState', DocumentState);
-}
